fixed relative path in import statements. moved parse kwargs utility method
